Separated bootstrapper, fixed the button hitbox and added sounds for menu buttons and timer
