+++ Assets/Graphics/HUD/ClickerButton.png.meta
@@ -21,7 +21,7 @@ TextureImporter:
     heightScale: 0.25
     normalMapFilter: 0
     flipGreenChannel: 0
-  isReadable: 0
+  isReadable: 1
   streamingMipmaps: 0
   streamingMipmapsPriority: 0
   vTOnly: 0
@@ -53,7 +53,7 @@ TextureImporter:
   spriteGenerateFallbackPhysicsShape: 1
   alphaUsage: 1
   alphaIsTransparency: 1
-  spriteTessellationDetail: -1
+  spriteTessellationDetail: 0
   textureType: 8
   textureShape: 1
   singleChannelComponent: 0
@@ -73,7 +73,7 @@ TextureImporter:
     resizeAlgorithm: 0
     textureFormat: -1
     textureCompression: 1
-    compressionQuality: 50
+    compressionQuality: 0
     crunchedCompression: 0
     allowsAlphaSplitting: 0
     overridden: 0
@@ -122,11 +122,56 @@ TextureImporter:
   spriteSheet:
     serializedVersion: 2
     sprites: []
-    outline: []
-    physicsShape: []
+    outline:
+    - - {x: 17, y: 64}
+      - {x: -3, y: 66}
+      - {x: -18, y: 63}
+      - {x: -41, y: 52}
+      - {x: -56, y: 36}
+      - {x: -66, y: 8}
+      - {x: -64, y: -17}
+      - {x: -56, y: -33}
+      - {x: -46, y: -48}
+      - {x: -29, y: -59}
+      - {x: -13, y: -64}
+      - {x: 4, y: -65}
+      - {x: 19, y: -64}
+      - {x: 36, y: -55}
+      - {x: 51, y: -41}
+      - {x: 62, y: -23}
+      - {x: 65, y: -3}
+      - {x: 64, y: 16}
+      - {x: 55, y: 38}
+      - {x: 36, y: 56}
+    physicsShape:
+    - - {x: 50, y: 44}
+      - {x: 32, y: 57}
+      - {x: 19, y: 63}
+      - {x: 3, y: 65}
+      - {x: -9, y: 65}
+      - {x: -27, y: 60}
+      - {x: -40, y: 52}
+      - {x: -50, y: 43}
+      - {x: -59, y: 30}
+      - {x: -65, y: 13}
+      - {x: -66, y: -3}
+      - {x: -64, y: -17}
+      - {x: -58, y: -32}
+      - {x: -47, y: -45}
+      - {x: -33, y: -56}
+      - {x: -17, y: -64}
+      - {x: 2, y: -65}
+      - {x: 15, y: -65}
+      - {x: 35, y: -56}
+      - {x: 48, y: -44}
+      - {x: 58, y: -31}
+      - {x: 64, y: -15}
+      - {x: 65, y: -1}
+      - {x: 65, y: 13}
+      - {x: 61, y: 25}
     bones: []
     spriteID: 5e97eb03825dee720800000000000000
-    internalID: 0
+    internalID: 1537655665
     vertices: []
     indices: 
     edges: []
