+++ Assets/Scripts/Bootstrapper.cs
@@ -7,36 +7,24 @@ using TMPro;
 using UnityEditor;
 using UnityEngine;
 using UnityEngine.Serialization;
-using UnityEngine.UI;
 using UnityEngine.InputSystem;
 using UnityEngine.EventSystems;
 #endregion
 public class Bootstrapper : MonoBehaviour
 {
     #region Serialized variables
-    [SerializeField] private int monthDuration;
-
-    [SerializeField] private bool IsRadio;
-
     [SerializeField] private TextMeshProUGUI resourcesLabel;
     [SerializeField] private TextMeshProUGUI essentialsLabel;
     [SerializeField] private TextMeshProUGUI peopleLabel;
 
-    [SerializeField] private TextMeshProUGUI yearLabel;
-    [SerializeField] private TextMeshProUGUI timeLabel;
-
-    [SerializeField] private AudioSource audioSource;
-    [SerializeField] private AudioClip[] AmbienceClips;
-    [SerializeField] private AudioClip[] RadioClips;
+    [SerializeField] private AudioSource sfxAudioSource;
+    [SerializeField] private AudioClip[] sfxClips;
     #endregion
 
     #region Serialized to static variables
     private static TextMeshProUGUI _resourcesLabel;
     private static TextMeshProUGUI _essentialsLabel;
     private static TextMeshProUGUI _peopleLabel;
-    
-    private static TextMeshProUGUI _yearLabel;
-    private static TextMeshProUGUI _timeLabel;
     #endregion
 
     #region Normal variables
@@ -46,7 +34,6 @@ public class Bootstrapper : MonoBehaviour
     #endregion
 
     #region Object variables
-    private TimeBank _timeObject;
     private ResourceBank _resourceObject;
     #endregion
 
@@ -56,9 +43,6 @@ public class Bootstrapper : MonoBehaviour
         _resourcesLabel = resourcesLabel;
         _essentialsLabel = essentialsLabel;
         _peopleLabel = peopleLabel;
-
-        _yearLabel = yearLabel;
-        _timeLabel = timeLabel;
         #endregion
 
         #region Normal realization
@@ -69,35 +53,17 @@ public class Bootstrapper : MonoBehaviour
 
         #region Object realization
         _resourceObject = new ResourceBank();
-        _timeObject = new TimeBank();
-        #endregion
-
-        #region Play music
-        audioSource.clip = GetRandomClip();
-        audioSource.Play();
         #endregion
 
         #region Invoke events
-        Invoke(nameof(EventOnEnd), audioSource.clip.length);
-
         _controlAsset.Clicker.Enable();
         #endregion
-
-        #region Start coroutines
-        StartCoroutine(TimeCoroutine());
-        #endregion
     }
+
     #region Normal functions
-    private AudioClip GetRandomClip()
+    private AudioClip GetSFXClips()
     {
-        if (IsRadio)
-        {
-            return RadioClips[UnityEngine.Random.Range(0, RadioClips.Length)];
-        }
-        else
-        {
-            return AmbienceClips[UnityEngine.Random.Range(0, AmbienceClips.Length)];
-        }
+        return sfxClips[UnityEngine.Random.Range(0, sfxClips.Length)];
     }
     #endregion
 
@@ -105,30 +71,11 @@ public class Bootstrapper : MonoBehaviour
     public void OnButtonClick()
     {
         _resourceObject.Resources += _clickPower;
-    }
-
-    void EventOnEnd()
-    {
-        audioSource.clip = GetRandomClip();
-        audioSource.Play();
+        sfxAudioSource.clip = GetSFXClips();
+        sfxAudioSource.Play();
     }
     #endregion
 
-    #region Coroutines
-    private IEnumerator TimeCoroutine()
-    {
-        while(true)
-        {
-            yield return new WaitForSeconds(monthDuration);
-            _timeObject.Month += 1;
-            if (_timeObject.Month > 12)
-            {
-                _timeObject.Month = 1;
-                _timeObject.Year += 1;
-            }
-        }
-    }
-    #endregion
     private class ResourceBank
     {
         #region Fields
@@ -163,35 +110,6 @@ public class Bootstrapper : MonoBehaviour
             {
                 _people = value;
                 _peopleLabel.text = $"Essentials: {_people}";
-            }
-        }
-        #endregion
-    }
-    
-    private class TimeBank
-    {
-        #region Fields
-        private int _year;
-        private int _month = 1;
-        #endregion
-
-        #region Properties
-        public int Year
-        {
-            get => _year;
-            set
-            {
-                _year = value;
-                _yearLabel.text = $"Year: {_year}";
-            }
-        }
-        public int Month
-        {
-            get => _month;
-            set
-            {
-                _month = value;
-                _timeLabel.text = $"Month: {Mathf.Round(_month)}";
             }
         }
         #endregion
