+++ Assets/Scripts/Music.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c0acf1d902322446a1641b8a98b9549
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
