+++ Assets/Scripts/Timer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b4778dd9290c7a249b2b63e6310ec167
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
